+++ src/avt/Pipeline/Data/avtPointExtractor.C
@@ -76,12 +76,12 @@ avtPointExtractor::avtPointExtractor(int w, int h, int d,
     // away a zone can ever be from samples).  Whatever this distance is, we
     // want R-max to be 50% bigger.  So calculate this by calculating the
     // position of something between the 8 samples ([0|1], [0|1], [0|1]).
-    float xS = (0*2. / width) - 1.;   // = -1.
-    float yS = (0*2. / height) - 1.;  // = -1.
-    float zS = 0.;
-    float xZ = (0.5*2. / width) - 1.;
-    float yZ = (0.5*2. / height) - 1.;
-    float zZ = 1. / depth;
+    double xS = (0*2. / width) - 1.;   // = -1.
+    double yS = (0*2. / height) - 1.;  // = -1.
+    double zS = 0.;
+    double xZ = (0.5*2. / width) - 1.;
+    double yZ = (0.5*2. / height) - 1.;
+    double zZ = 1. / depth;
     smallestX = 1.5*(xZ - xS);
     smallestY = 1.5*(yZ - yS);
     smallestZ = 1.5*(zZ - zS);
@@ -128,20 +128,20 @@ avtPointExtractor::~avtPointExtractor()
 void
 avtPointExtractor::Extract(const avtPoint &pt)
 {
-    const float *bbox = pt.bbox;
+    const double *bbox = pt.bbox;
 
-    float center[3];
+    double center[3];
     center[0] = (bbox[0] + bbox[1]) / 2.;
     center[1] = (bbox[2] + bbox[3]) / 2.;
     center[2] = (bbox[4] + bbox[5]) / 2.;
-    float xDiff = (center[0] - bbox[0]);
+    double xDiff = (center[0] - bbox[0]);
     xDiff = (xDiff > smallestX ? xDiff : smallestX);
-    float yDiff = (center[1] - bbox[2]);
+    double yDiff = (center[1] - bbox[2]);
     yDiff = (yDiff > smallestY ? yDiff : smallestY);
-    float zDiff = (center[2] - bbox[4]);
+    double zDiff = (center[2] - bbox[4]);
     zDiff = (zDiff > smallestZ ? zDiff : smallestZ);
 
-    float rad = 1.5*sqrt(xDiff*xDiff + yDiff*yDiff + zDiff*zDiff);
+    double rad = 1.5*sqrt(xDiff*xDiff + yDiff*yDiff + zDiff*zDiff);
     int minx = (int) ceil((((center[0] - rad) + 1.) / 2.) * width);
     int maxx = (int) floor((((center[0] + rad) + 1.) / 2.) * width);
     int miny = (int) ceil((((center[1] - rad) + 1.) / 2.) * height);
@@ -177,29 +177,29 @@ avtPointExtractor::Extract(const avtPoint &pt)
         return;
     }
 
-    float cval2[AVT_VARIABLE_LIMIT];
+    double cval2[AVT_VARIABLE_LIMIT];
     int nVars = pt.nVars;
 
     for (int i = minx ; i <= maxx ; i++)
     {
         if (i < 0 || i >= width)
             continue;
-        float xf = ((i*2.) / width)  - 1.;
-        float xD = xf-center[0];
+        double xf = ((i*2.) / width)  - 1.;
+        double xD = xf-center[0];
         for (int j = miny ; j <= maxy ; j++)
         {
             if (j < 0 || j >= height)
                 continue;
             avtRay *ray = volume->GetRay(i, j);
-            float yf = ((j*2.) / height) - 1.;
-            float yD = yf-center[1];
+            double yf = ((j*2.) / height) - 1.;
+            double yD = yf-center[1];
             for (int k = minz ; k <= maxz ; k++)
             {
                 if (k < 0 || k >= depth)
                     continue;
-                //float zf = ((k*2.) / depth)  - 1.;
-                float zf = ((float)k)/((float)depth);
-                float zD = 2*(zf-center[2]);
+                //double zf = ((k*2.) / depth)  - 1.;
+                double zf = ((double)k)/((double)depth);
+                double zD = 2*(zf-center[2]);
                 double dist = xD*xD + yD*yD + zD*zD;
                 double weight = 0.;
                 if (dist >= rad*rad)
